Updates to add_product
- Add button to link from dashboard products page to add product form
- Made base of add_product form the dashboard base
- Move forms from Inventory forms.py to Orders forms.py
- Refactored jquery in add_product.html
- Mistake fixed in products.html where price wasn't coming in for Other Products
