+++ inventory/forms.py
@@ -71,46 +71,3 @@ class InventoryKillCropForm(forms.Form):
         super(InventoryKillCropForm, self).__init__(*args, **kwargs)
         for reason in KillReason.objects.all():
             self.fields[reason.name] = forms.BooleanField()
-
-
-class AddProductForm(forms.Form):
-    product_name = forms.CharField(max_length=200)
-    price = forms.FloatField(min_value=0.0)
-
-    def get_live_crop_fields(self):
-        variety = forms.ModelChoiceField(queryset=Variety.objects.all(),
-                                         widget=forms.Select(attrs={'class': 'form-control'}))
-        size = forms.ModelChoiceField(queryset=MicrogreenSize.objects.all(),
-                                      widget=forms.Select(attrs={'class': 'form-control'}))
-        tray_type = forms.ModelChoiceField(queryset=TrayType.objects.all(),
-                                           widget=forms.Select(attrs={'class': 'form-control'}))
-        return [variety, size, tray_type]
-
-    def get_harvested_crop_fields(self):
-        variety = forms.ModelChoiceField(queryset=Variety.objects.all(),
-                                         widget=forms.Select(attrs={'class': 'form-control'}))
-        size = forms.ModelChoiceField(queryset=MicrogreenSize.objects.all(),
-                                      widget=forms.Select(attrs={'class': 'form-control'}))
-        weight = forms.FloatField(min_value=0.0)
-        return [variety, size, weight]
-
-
-class AddLiveCropProductForm(forms.Form):
-    product_name = forms.CharField(max_length=200)
-    price = forms.FloatField(min_value=0.0)
-    variety = forms.ModelChoiceField(queryset=Variety.objects.all(),
-                                     widget=forms.Select(attrs={'class': 'form-control'}))
-    size = forms.ModelChoiceField(queryset=MicrogreenSize.objects.all(),
-                                  widget=forms.Select(attrs={'class': 'form-control'}))
-    tray_type = forms.ModelChoiceField(queryset=TrayType.objects.all(),
-                                  widget=forms.Select(attrs={'class': 'form-control'}))
-
-
-class AddHarvestCropProductForm(forms.Form):
-    product_name = forms.CharField(max_length=200)
-    price = forms.FloatField(min_value=0.0)
-    variety = forms.ModelChoiceField(queryset=Variety.objects.all(),
-                                     widget=forms.Select(attrs={'class': 'form-control'}))
-    size = forms.ModelChoiceField(queryset=MicrogreenSize.objects.all(),
-                                  widget=forms.Select(attrs={'class': 'form-control'}))
-    weight = forms.FloatField(min_value=0.0)
